+++ mrmustard/math/lattice/strategies/compactFock/compactFock_inputValidation.py
@@ -6,12 +6,12 @@ to validate the input provided by the user.
 
 from typing import Iterable
 import numpy as np
-from mrmustard.math.lattice.strategies.compactFock.python.compactFock_diagonal_amps import fock_representation_diagonal_amps
-from mrmustard.math.lattice.strategies.compactFock.python.compactFock_diagonal_grad import fock_representation_diagonal_grad
-from mrmustard.math.lattice.strategies.compactFock.python.compactFock_1leftoverMode_amps import (
+from mrmustard.math.lattice.strategies.compactFock.compactFock_diagonal_amps import fock_representation_diagonal_amps
+from mrmustard.math.lattice.strategies.compactFock.compactFock_diagonal_grad import fock_representation_diagonal_grad
+from mrmustard.math.lattice.strategies.compactFock.compactFock_1leftoverMode_amps import (
     fock_representation_1leftoverMode_amps,
 )
-from mrmustard.math.lattice.strategies.compactFock.python.compactFock_1leftoverMode_grad import (
+from mrmustard.math.lattice.strategies.compactFock.compactFock_1leftoverMode_grad import (
     fock_representation_1leftoverMode_grad,
 )
 from thewalrus._hafnian import input_validation
