+++ mrmustard/math/tensorflow_wrapper.py
@@ -23,7 +23,7 @@ import tensorflow_probability as tfp
 from mrmustard import settings
 from mrmustard.math.autocast import Autocast
 from mrmustard.math.lattice import strategies
-from mrmustard.math.lattice.strategies.compactFock.python.compactFock_inputValidation import (
+from mrmustard.math.lattice.strategies.compactFock.compactFock_inputValidation import (
     grad_hermite_multidimensional_1leftoverMode,
     grad_hermite_multidimensional_diagonal,
     hermite_multidimensional_1leftoverMode,
@@ -398,7 +398,7 @@ class TFMath(MathInterface):
                 _B.astype(np.complex128),
                 _C.astype(np.complex128),
             )
-            G = Main_julia.vanilla(_A, _B, _C.item(), np.array(shape, dtype=np.int64))
+            G = Main_julia.Vanilla.vanilla(_A, _B, _C.item(), np.array(shape, dtype=np.int64))
 
         def grad(dLdGconj):
             dLdA, dLdB, dLdC = strategies.vanilla_vjp(G, _C, np.conj(dLdGconj))
