+++ mrmustard/settings.py
@@ -301,7 +301,13 @@ class Settings:
             # import Julia functions
             math_directory = os.path.dirname(__file__)
             Main_julia.cd(math_directory)
-            Main_julia.include("math/lattice/strategies/vanilla.jl")
+            Main_julia.include("math/lattice/strategies/julia/helperFunctions.jl")
+            Main_julia.include("math/lattice/strategies/julia/vanilla.jl")
+            Main_julia.include("math/lattice/strategies/julia/compactFock/compactFock_helperFunctions.jl")
+            Main_julia.include("math/lattice/strategies/julia/compactFock/diagonal_amps.jl")
+            Main_julia.include("math/lattice/strategies/julia/compactFock/diagonal_grad.jl")
+            Main_julia.include("math/lattice/strategies/julia/compactFock/1leftoverMode_amps.jl")
+            Main_julia.include("math/lattice/strategies/julia/compactFock/1leftoverMode_grad.jl")
 
     # use rich.table to print the settings
     def __repr__(self) -> str:
